fix: forge format

+++ src/CustomHook.sol
@@ -20,27 +20,16 @@ contract CustomHook is BaseHook, Ownable {
     using BeforeSwapDeltaLibrary for BeforeSwapDelta;
     using SafeERC20 for IERC20;
 
-    Currency public immutable USDC =
-        Currency.wrap(0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48);
-    Currency public immutable USDT =
-        Currency.wrap(0xdAC17F958D2ee523a2206206994597C13D831ec7);
+    Currency public immutable USDC = Currency.wrap(0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48);
+    Currency public immutable USDT = Currency.wrap(0xdAC17F958D2ee523a2206206994597C13D831ec7);
 
     error InvalidPoolTokens();
 
-    constructor(
-        IPoolManager _poolManager,
-        IAavePool _aavePool
-    ) BaseHook(_poolManager) Ownable(msg.sender) {
+    constructor(IPoolManager _poolManager, IAavePool _aavePool) BaseHook(_poolManager) Ownable(msg.sender) {
         aavePool = _aavePool;
 
-        IERC20(Currency.unwrap(USDC)).forceApprove(
-            address(aavePool),
-            type(uint).max
-        );
-        IERC20(Currency.unwrap(USDT)).forceApprove(
-            address(aavePool),
-            type(uint).max
-        );
+        IERC20(Currency.unwrap(USDC)).forceApprove(address(aavePool), type(uint256).max);
+        IERC20(Currency.unwrap(USDT)).forceApprove(address(aavePool), type(uint256).max);
     }
 
     // -----------------------------------------------
@@ -52,39 +41,29 @@ contract CustomHook is BaseHook, Ownable {
     /// - beforeInitialize
     /// - beforeSwap
     /// - beforeSwapReturnDelta
-    function getHookPermissions()
-        public
-        pure
-        override
-        returns (Hooks.Permissions memory)
-    {
-        return
-            Hooks.Permissions({
-                beforeInitialize: true,
-                afterInitialize: false,
-                beforeAddLiquidity: false,
-                afterAddLiquidity: false,
-                beforeRemoveLiquidity: false,
-                afterRemoveLiquidity: false,
-                beforeSwap: true,
-                afterSwap: false,
-                beforeDonate: false,
-                afterDonate: false,
-                beforeSwapReturnDelta: true,
-                afterSwapReturnDelta: false,
-                afterAddLiquidityReturnDelta: false,
-                afterRemoveLiquidityReturnDelta: false
-            });
+    function getHookPermissions() public pure override returns (Hooks.Permissions memory) {
+        return Hooks.Permissions({
+            beforeInitialize: true,
+            afterInitialize: false,
+            beforeAddLiquidity: false,
+            afterAddLiquidity: false,
+            beforeRemoveLiquidity: false,
+            afterRemoveLiquidity: false,
+            beforeSwap: true,
+            afterSwap: false,
+            beforeDonate: false,
+            afterDonate: false,
+            beforeSwapReturnDelta: true,
+            afterSwapReturnDelta: false,
+            afterAddLiquidityReturnDelta: false,
+            afterRemoveLiquidityReturnDelta: false
+        });
     }
 
     /// @notice Inherit from _beforeInitialize of BaseHook
     /// Implemented logic:
     /// pool token validation in beforeInitialize hook
-    function _beforeInitialize(
-        address,
-        PoolKey calldata key,
-        uint160
-    ) internal override returns (bytes4) {
+    function _beforeInitialize(address, PoolKey calldata key, uint160) internal override returns (bytes4) {
         if (key.currency0 == USDC && key.currency1 == USDT) {
             return BaseHook.beforeInitialize.selector;
         }
@@ -105,15 +84,12 @@ contract CustomHook is BaseHook, Ownable {
         IPoolManager.SwapParams calldata swapParams,
         bytes calldata hookData
     ) internal override returns (bytes4, BeforeSwapDelta, uint24) {
-        uint swapAmount = uint(
-            swapParams.amountSpecified > 0
-                ? swapParams.amountSpecified
-                : -swapParams.amountSpecified
-        );
+        uint256 swapAmount =
+            uint256(swapParams.amountSpecified > 0 ? swapParams.amountSpecified : -swapParams.amountSpecified);
         Currency swapCurrency = swapParams.zeroForOne ? USDC : USDT;
 
         // calculate fee
-        uint fee = swapAmount / 1000;
+        uint256 fee = swapAmount / 1000;
         totalFeeAccrued[swapCurrency] += fee;
         devFeeAccrued[swapCurrency] += fee / 2;
         poolManager.take(swapCurrency, address(this), fee);
@@ -137,14 +113,15 @@ contract CustomHook is BaseHook, Ownable {
     // -----------------------------------------------
 
     struct UserInfo {
-        uint amount;
-        uint rewardDebt;
+        uint256 amount;
+        uint256 rewardDebt;
     }
+
     mapping(Currency => mapping(address => UserInfo)) public userInfos;
-    mapping(Currency => uint) public devFeeAccrued;
-    mapping(Currency => uint) public totalFeeAccrued;
-    mapping(Currency => uint) public rewardPerShare;
-    mapping(Currency => uint) public totalShares;
+    mapping(Currency => uint256) public devFeeAccrued;
+    mapping(Currency => uint256) public totalFeeAccrued;
+    mapping(Currency => uint256) public rewardPerShare;
+    mapping(Currency => uint256) public totalShares;
 
     /// @notice Encode hook data
     function getHookData(address user) public pure returns (bytes memory) {
@@ -152,34 +129,24 @@ contract CustomHook is BaseHook, Ownable {
     }
 
     /// @notice Decode hook data
-    function parseHookData(
-        bytes calldata data
-    ) public pure returns (address user) {
+    function parseHookData(bytes calldata data) public pure returns (address user) {
         return abi.decode(data, (address));
     }
 
     /// @notice Pending rewards amount in virtual masterchef
-    function pendingRewards(
-        address user,
-        Currency currency
-    ) public view returns (uint reward) {
+    function pendingRewards(address user, Currency currency) public view returns (uint256 reward) {
         UserInfo storage userInfo = userInfos[currency][user];
-        reward =
-            (userInfo.amount * rewardPerShare[currency]) /
-            1e18 -
-            userInfo.rewardDebt;
+        reward = (userInfo.amount * rewardPerShare[currency]) / 1e18 - userInfo.rewardDebt;
     }
 
     /// @notice Claim user fee, called by anyone
     /// Withdraw fee from Aave pool and returns to user
     function claimUserFee(Currency currency) external {
         address user = msg.sender;
-        uint pendingReward = pendingRewards(user, currency);
+        uint256 pendingReward = pendingRewards(user, currency);
         if (pendingReward > 0) {
             UserInfo storage userInfo = userInfos[currency][user];
-            userInfo.rewardDebt =
-                (userInfo.amount * rewardPerShare[currency]) /
-                1e18;
+            userInfo.rewardDebt = (userInfo.amount * rewardPerShare[currency]) / 1e18;
             _withdrawFromAave(pendingReward, user, currency);
         }
     }
@@ -193,11 +160,7 @@ contract CustomHook is BaseHook, Ownable {
 
     /// @notice Deposit to virtual masterchef
     /// This function is needed for reward calculation
-    function _depositToVirtualMC(
-        address user,
-        uint amount,
-        Currency currency
-    ) internal {
+    function _depositToVirtualMC(address user, uint256 amount, Currency currency) internal {
         UserInfo storage userInfo = userInfos[currency][user];
         userInfo.amount += amount;
         userInfo.rewardDebt += (amount * rewardPerShare[currency]) / 1e18;
@@ -205,7 +168,7 @@ contract CustomHook is BaseHook, Ownable {
     }
 
     /// @notice Add reward to virtual masterchef
-    function _addRewardsToVirtualMC(Currency currency, uint amount) internal {
+    function _addRewardsToVirtualMC(Currency currency, uint256 amount) internal {
         rewardPerShare[currency] += (amount * 1e18) / totalShares[currency];
     }
 
@@ -214,7 +177,7 @@ contract CustomHook is BaseHook, Ownable {
     // -----------------------------------------------
 
     mapping(Currency => IERC20) public aTokens;
-    mapping(Currency => uint) public totalATokenShares;
+    mapping(Currency => uint256) public totalATokenShares;
     IAavePool public immutable aavePool;
 
     /// @notice Set aToken of currency
@@ -223,27 +186,19 @@ contract CustomHook is BaseHook, Ownable {
     }
 
     /// @notice Aave's aToken share price calcuation
-    function _getATokenSharePrice(
-        Currency currency
-    ) internal view returns (uint) {
-        return
-            (aTokens[currency].balanceOf(address(this)) * 1e18) /
-            totalATokenShares[currency];
+    function _getATokenSharePrice(Currency currency) internal view returns (uint256) {
+        return (aTokens[currency].balanceOf(address(this)) * 1e18) / totalATokenShares[currency];
     }
 
     /// @notice Deposit to aave pool
-    function _depositToAave(uint share, Currency currency) internal {
+    function _depositToAave(uint256 share, Currency currency) internal {
         totalATokenShares[currency] += share;
         aavePool.supply(Currency.unwrap(currency), share, address(this), 0);
     }
 
     /// @notice Withdraw from aave pool
-    function _withdrawFromAave(
-        uint share,
-        address to,
-        Currency currency
-    ) internal returns (uint) {
-        uint amount = (share * _getATokenSharePrice(currency)) / 1e18;
+    function _withdrawFromAave(uint256 share, address to, Currency currency) internal returns (uint256) {
+        uint256 amount = (share * _getATokenSharePrice(currency)) / 1e18;
         totalATokenShares[currency] -= share;
         return aavePool.withdraw(Currency.unwrap(currency), amount, to);
     }
